Tests: Sleep: Extract test utility functions
Move a few of utility functions so other tests can use them.
